+++ Marlin/Marlin/UI/Port/Sheet/PortSheetView.swift
@@ -26,7 +26,7 @@ struct PortSheetView: View {
             }
         }
         .onChange(of: itemKey) { _ in
-            viewModel.getPort(portNumber: Int64(itemKey) ?? -1)
+            viewModel.getPort(portNumber: Int(itemKey) ?? -1)
         }
         .onChange(of: viewModel.port) { model in
             NotificationCenter.default.post(
@@ -39,7 +39,7 @@ struct PortSheetView: View {
         }
         .onAppear {
             viewModel.repository = portRepository
-            viewModel.getPort(portNumber: Int64(itemKey) ?? -1)
+            viewModel.getPort(portNumber: Int(itemKey) ?? -1)
         }
     }
 }
